Dhananjay singh 21_03_2025 4pm
Dhananjay singh 21_03_2025 4:12pm

+++ app/src/main/java/com/mykaimeal/planner/fragment/mainfragment/commonscreen/basketscreen/model/BasketScreenModel.kt
@@ -58,12 +58,11 @@ data class Address(
 
 data class BillingAddress(
     val recipes:Int?,
-    val net_total:Int?,
-    val tax:Int?,
-    val delivery:Int?,
+    val net_total:Double?,
+    val tax:Double?,
+    val delivery:Double?,
     val processing:Double?,
-    val total:Int?
-
+    val total:Double?
 )
